requirements.txt
openai>=1.0.0
google-api-python-client>=2.0.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=1.0.0
gspread>=5.0.0
python-dateutil>=2.8.0
pytz>=2023.0
PyPDF2>=3.0.0
python-dotenv>=1.0.0
